Fixed rating; disabled old caching methods; using taxon_concept_content to cache best image

+++ app/models/taxon_concept_exemplar_image.rb
@@ -3,13 +3,18 @@ class TaxonConceptExemplarImage < SpeciesSchemaModel
   has_many :taxon_concepts
   set_primary_key :taxon_concept_id
 
-  def self.set_exemplar(taxon_concept_id, data_object_id)
-    exemplar = self.find_or_create_by_taxon_concept_id(taxon_concept_id)
+  def self.set_exemplar(taxon_concept, data_object_id)
+    return if taxon_concept.nil?
+    exemplar = self.find_or_create_by_taxon_concept_id(taxon_concept.id)
     exemplar.update_attribute(:data_object_id, data_object_id)
 
-    tci_exists = TopConceptImage.find_by_taxon_concept_id_and_data_object_id(taxon_concept_id, data_object_id)
+    tci_exists = TopConceptImage.find_by_taxon_concept_id_and_data_object_id(taxon_concept.id, data_object_id)
     tci_exists.destroy unless tci_exists.nil?
-    connection.execute("UPDATE top_concept_images SET view_order=view_order+1 WHERE taxon_concept_id=#{taxon_concept_id}");
-    TopConceptImage.create(:taxon_concept_id => taxon_concept_id, :data_object_id => data_object_id, :view_order => 1)
+    connection.execute("UPDATE top_concept_images SET view_order=view_order+1 WHERE taxon_concept_id=#{taxon_concept.id}");
+    TopConceptImage.create(:taxon_concept_id => taxon_concept.id, :data_object_id => data_object_id, :view_order => 1)
+    if taxon_concept.taxon_concept_content
+      taxon_concept.taxon_concept_content.image_object_id = data_object_id
+      taxon_concept.taxon_concept_content.save
+    end
   end
 end
